This is the script
It's the only file in this repo.  Edit the variables at the top to change the directory on both the target and source, as well as any connection info.  I used it to move media files from a raspberry pi to a windows machine, so there are a few steps taken to consider typical problems with sending large files over sftp.